fully locked down the back-end in monitoring mode

+++ includes/hostapd.php
@@ -2,6 +2,7 @@
 
 include_once('includes/status_messages.php');
 include_once('app/lib/system.php');
+require_once 'config.php';
 
 /**
 *
@@ -25,23 +26,25 @@ function DisplayHostAPDConfig()
     $managedModeEnabled = false;
     exec("ip -o link show | awk -F': ' '{print $2}'", $interfaces);
 
-    if (isset($_POST['SaveHostAPDSettings'])) {
-        SaveHostAPDConfig($arrSecurity, $arrEncType, $arr80211Standard, $interfaces, $status);
-    } elseif (isset($_POST['StartHotspot'])) {
-        $status->addMessage('Attempting to start hotspot', 'info');
-        if ($arrHostapdConf['WifiAPEnable'] == 1) {
-            exec('sudo /etc/raspap/hostapd/servicestart.sh --interface uap0 --seconds 3', $return);
-        } else {
-            exec('sudo /etc/raspap/hostapd/servicestart.sh --seconds 5', $return);
-        }
-        foreach ($return as $line) {
-            $status->addMessage($line, 'info');
-        }
-    } elseif (isset($_POST['StopHotspot'])) {
-        $status->addMessage('Attempting to stop hotspot', 'info');
-        exec('sudo /bin/systemctl stop hostapd.service', $return);
-        foreach ($return as $line) {
-            $status->addMessage($line, 'info');
+    if (!RASPI_MONITOR_ENABLED) {
+        if (isset($_POST['SaveHostAPDSettings'])) {
+            SaveHostAPDConfig($arrSecurity, $arrEncType, $arr80211Standard, $interfaces, $status);
+        } elseif (isset($_POST['StartHotspot'])) {
+            $status->addMessage('Attempting to start hotspot', 'info');
+            if ($arrHostapdConf['WifiAPEnable'] == 1) {
+                exec('sudo /etc/raspap/hostapd/servicestart.sh --interface uap0 --seconds 3', $return);
+            } else {
+                exec('sudo /etc/raspap/hostapd/servicestart.sh --seconds 5', $return);
+            }
+            foreach ($return as $line) {
+                $status->addMessage($line, 'info');
+            }
+        } elseif (isset($_POST['StopHotspot'])) {
+            $status->addMessage('Attempting to stop hotspot', 'info');
+            exec('sudo /bin/systemctl stop hostapd.service', $return);
+            foreach ($return as $line) {
+                $status->addMessage($line, 'info');
+            }
         }
     }
